[move-prover] Removing AbstractType because it is not sound.
A while ago we introduced a special type constructor `AbstractType` which we assigned to type parameters for top-level entry points. This was meant to make extensional equality of types working, and also was a candidate for the performance problem with Z3 we observeed. However, this is unsound because then Z3 can prove that R<u64> is not equal R<T> for any type paramter T. This PR removes AbstractType and adds a regressison test to ensures this continues to work.

However, with this PR it's again unclear whether we can assume equality to work for type values, which we need because we index global memory with it. For this, we may need to come up with a different solution.

Closes: #3952
Approved by: shazqadeer