[network][connmgr] Improve logging of peer address updates
Recently, while debugging some connectivity issues, it was noticably
difficult to glean the current state of the discovery info in connectivity
manager.

This diff is a first step in trying to improve network debuggability by
ensuring that we actually log updates to the peer_addresses set (with
some small tweaks so we only log when it actually updates).

Closes: #4353
Approved by: davidiw