[config] Remove seed and network peer files
These files may be useful at some point in time, but I would argue they
aren't particularly useful at this point in time and add unnecessary
complexity to the configs.

There's more work we could do here, but let's delay.

Note the tests are only really there to verify storing and loading the
configs. So with them gone, we're good to go.

Also it looks like the generation of network address test was broken. At
this point in time, the only test really for this code is the
config-builder / smoke tests

Closes: #4465
Approved by: gregnazario