[move framework] Simplify LibraConfig for verification
Incrementally make LibraConfig permissions depend on whether
signer is Libra root instead of capabilities.

This retains the use of the basic ModifyConfigCapability, which is
stored at the address of the signer who is authoried to modify the
capability (currently, that is always libra root).

Updated LibraConfig and a bunch of files that use configs, plus a few
functional tests and LibraTest in docgen.

Closes: #5038
Approved by: sblackshear