[Libra Framework][Specs] Use global invariants where appropriate.
This PR converts existing invariants in the system into the new global invariants (`invariant [global] P` or `invariant update [global] P`).

Global invariants are (at least currently) faster than module invariants because they are only asserted when memory they are depending on is updated. Similarily, they are assumed before memory is read. This is up to the level of granularity of the memory of a given resource type; where the update in the memory happens is currently not differentiated.

Global invariants are more expressive, because they can express interdependency of resource states, but this PR just converts existing oens. As a guide for porting existing invariants, the following holds:

- Where before we wrote `spec module { invariant P }` we now write `invariant [global] P`.
- Where before we wrote `schema S { ensures P; } spec module { apply S to *; }` we now write `spec module { invariant update [global] P; }`.

This change brings some moderate performance improvements, and disabled verification could be turned on. There is one case left in DesignatedDealer which does verify only after a long time, and one case in LibraAccount which never terminates. However, there is no 100% evidence that everything is sound, because the feature is still new.

Closes: #5381
Approved by: junkil-park