[move-prover] Restore global timeout flag and leverage new boogie features for per-procedure settings.
This restores the temporarily broken `--timeout/-T` flag of the prover and leverages new boogie features to set seed and timeout via a praggma. For example:

```
spec fun foo {
    pragma seed = 928702;
    pragma timeout = 100 /*seconds*/;
}
```

The PR also makes us ready for Ci by tweaking timeouts to have an added 50% buffer if running on a Linux flavor, like in Ci. This should help to prevent flakiness.

Closes: #5387
Approved by: shazqadeer