[cluster-test] add spawn_job and kill_job for ClusterSwarmKube
This allows us to spawn long-running jobs like faucet, execute
depedent jobs such as for CLI commands, and kill it later on.

Adds a skeleton for client_compatibility_test, which will include
spawning a faucet as a k8s job, running CLI jobs to mint against
that, and killing the faucet job at the end.

Closes: #5488