[genesis] expose the two key api from genesis encoding
Genesis actually expects a key libra root and one for treasury root but
since this was so new it wasn't exposed. From a testing perspective, it
is totally valid that they are the same key. Of course, this means that
the two different paths aren't as effectively tested, but ... this is
good enough for now.