[testsuite] remove compat and roar
Now that we've phased out the old ECS testnet, we can get rid of record
and replay, and also the ECS version of compat. Compat is already
running in cluster test. Record-and-replay implementation in k8s will be
a bit more involved. Corresponding changes need to be made on the infra side too.

Closes: #5900