[testsuite] Fixing CLI shorthands for "dev" subcommands
Adding shorthands for the "dev" subcommands still lacking one.

Since 0c870bff1 ("[cli] add consistency assertions for client_proxy
commands", 2020-04-13) deliberately adds consistency checks for some
of the subcommands, following that lead with shorthands too.

As a long term solution, removing the consistency checks and ensuring
proper operation in other ways could be the way to go in the future?

Closes: #6363