[tooling] Added a network checking tool for noise endpoints
Note, it's really hard to check if a noise endpoint is valid or not
without providing some dummy information that could then be exploited by
attackers.  This just does the due dilligence of checking if an address
is working, and that the endpoint shouldn't respond with anything even
if prompted.

Closes: #6415