[storage] arb_blocks_to_commit: rely NOT on the DB
We intend to make the DB dump, as in save whatever the upper layer asks to save, potentially losing the ability to calculate any hash. This removes the dependance on the DB's ability to calculate any Merkle hash, instead, the test helper being changed here relies on the in mem merkle tree algorithms directly.

Closes: #10159